+++ packages/react-native/jest.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  preset: 'react-native',
+  testMatch: ['<rootDir>/src/**/__tests__/**/*.test.ts'],
+  clearMocks: true,
+  restoreMocks: true,
+};